a few more windows warnings

+++ libraries/script-engine/src/v8/FastScriptValueUtils.cpp
@@ -56,7 +56,7 @@ bool qBytearrayFromScriptValue(const ScriptValue& object, QByteArray &qByteArray
         return false;
     }
     v8::Local<v8::ArrayBuffer> arrayBuffer = v8::Local<v8::ArrayBuffer>::Cast(v8Value);
-    qByteArray.resize(arrayBuffer->ByteLength());
+    qByteArray.resize((int)arrayBuffer->ByteLength());
     memcpy(qByteArray.data(), arrayBuffer->Data(), arrayBuffer->ByteLength());
     return true;
 }
